refactor: migrate state classes to package
